Create cpp file and add docs
